# Zero out protocol parameter instead of deleting to improve performance