This improves the performance by a 1000ms on a million runs.

Previously:

```sh
Run 0: Benchmark with no headers took: 2757.767ms
Run 1: Benchmark with no headers took: 2732.211ms
Run 2: Benchmark with no headers took: 2795.459ms
Run 3: Benchmark with no headers took: 2738.267ms
Run 0: Benchmark with Forwarded header took: 3436.794ms
Run 1: Benchmark with Forwarded header took: 3442.375ms
Run 2: Benchmark with Forwarded header took: 3413.114ms
Run 3: Benchmark with Forwarded header took: 3405.076ms
Run 0: Benchmark with multiple X-Forwarded-Host headers took: 2719.612ms
Run 1: Benchmark with multiple X-Forwarded-Host headers took: 2691.415ms
Run 2: Benchmark with multiple X-Forwarded-Host headers took: 2684.705ms
Run 3: Benchmark with multiple X-Forwarded-Host headers took: 2694.015ms
```

After:

```sh
Run 0: Benchmark with no headers took: 1759.048ms
Run 1: Benchmark with no headers took: 1741.105ms
Run 2: Benchmark with no headers took: 1733.430ms
Run 3: Benchmark with no headers took: 1766.368ms
Run 0: Benchmark with Forwarded header took: 2433.636ms
Run 1: Benchmark with Forwarded header took: 2409.549ms
Run 2: Benchmark with Forwarded header took: 2400.254ms
Run 3: Benchmark with Forwarded header took: 2413.262ms
Run 0: Benchmark with multiple X-Forwarded-Host headers took: 1785.926ms
Run 1: Benchmark with multiple X-Forwarded-Host headers took: 1786.500ms
Run 2: Benchmark with multiple X-Forwarded-Host headers took: 1783.425ms
Run 3: Benchmark with multiple X-Forwarded-Host headers took: 1787.137ms
```